+++ tests/e2e/matrix/internal/routes/admin-service/routes.go
@@ -4,7 +4,7 @@ package adminservice
 
 import (
 	"github.com/craftgodotdev/craftgo/pkg/server"
-	transport "github.com/craftgodotdev/craftgo/tests/e2e/matrix/internal/transport/admin-service"
+	transport "github.com/craftgodotdev/craftgo/tests/e2e/matrix/internal/transport/admin-service/ops/v2"
 	"github.com/craftgodotdev/craftgo/tests/e2e/matrix/svccontext"
 )